Добавляет компонент MainMenu.tsx
