+++ src/components/Drawer/Drawer.style.ts
@@ -1,0 +1,11 @@
+import { createStyles, makeStyles, Theme } from '@material-ui/core';
+
+const drawerWidth = 240;
+
+export default makeStyles((theme: Theme) =>
+  createStyles({
+    drawer: {
+      width: drawerWidth,
+    },
+  })
+);
